+++ test/integration/default-latest/serverspec/karaf_install_spec.rb
@@ -3,6 +3,6 @@ require 'karaf_install'
 set :backend, :exec
 
 describe 'karaf-latest' do
-  it_behaves_like 'karaf_install'
+  include_examples 'karaf_install'
 end