+++ test/integration/slow-cloud/serverspec/karaf_install_spec.rb
@@ -3,5 +3,5 @@ require 'karaf_install'
 set :backend, :exec
 
 describe 'slow-cloud' do
-  it_behaves_like 'karaf_install'
+  include_examples 'karaf_install'
 end